[WIP] Refactoring the schedulers
