+++ src/diffusers/__init__.py
@@ -31,11 +31,11 @@ if is_torch_available():
     from .pipeline_utils import DiffusionPipeline
     from .pipelines import DDIMPipeline, DDPMPipeline, KarrasVePipeline, LDMPipeline, PNDMPipeline, ScoreSdeVePipeline
     from .schedulers import (
+        BaseScheduler,
         DDIMScheduler,
         DDPMScheduler,
         KarrasVeScheduler,
         PNDMScheduler,
-        SchedulerMixin,
         ScoreSdeVeScheduler,
     )
     from .training_utils import EMAModel
