+++ src/diffusers/schedulers/__init__.py
@@ -23,7 +23,7 @@ if is_torch_available():
     from .scheduling_pndm import PNDMScheduler
     from .scheduling_sde_ve import ScoreSdeVeScheduler
     from .scheduling_sde_vp import ScoreSdeVpScheduler
-    from .scheduling_utils import SchedulerMixin
+    from .scheduling_utils import BaseScheduler
 else:
     from ..utils.dummy_pt_objects import *  # noqa F403
